handle non existant files for delete and head

+++ src/files_api/routes.py
@@ -100,6 +100,15 @@ async def get_file_metadata(
     Note: by convention, HEAD requests MUST NOT return a body in the response.
     """
     settings: Settings = request.app.state.settings
+
+    object_exists = object_exists_in_s3(
+        bucket_name=settings.s3_bucket_name, object_key=file_path
+    )
+    if not object_exists:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND, detail="File not found"
+        )
+
     get_object_response = fetch_s3_object(settings.s3_bucket_name, object_key=file_path)
     response.headers["Content-Type"] = get_object_response["ContentType"]
     response.headers["Content-Length"] = str(get_object_response["ContentLength"])
@@ -144,6 +153,14 @@ async def delete_file(
 
     NOTE: DELETE requests MUST NOT return a body in the response."""
     settings: Settings = request.app.state.settings
+
+    object_exists = object_exists_in_s3(
+        bucket_name=settings.s3_bucket_name, object_key=file_path
+    )
+    if not object_exists:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND, detail="File not found"
+        )
     delete_s3_object(settings.s3_bucket_name, object_key=file_path)
     response.status_code = status.HTTP_204_NO_CONTENT
     return response
